feat(front): improve admin routing with shared sidebar layout
